backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
flask_restful==0.3.9
axios
mysql-connector-python
bcrypt